refactor(core): extract Engine transcribe API wrappers
Move transcribe_file / transcribe_bytes* into engine/api.rs.
The request pipeline stays in transcribe.rs; public Engine
methods are unchanged.
